Removed the ID Token Hint from the PAT Grant.
