+++ components/org.wso2.carbon.identity.pat/org.wso2.carbon.identity.pat.api.rest.commons/src/main/java/org/wso2/carbon/identity/pat/api/rest/commons/Constants.java
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.com). All Rights Reserved.
+ *
+ * This software is the property of WSO2 Inc. and its suppliers, if any.
+ * Dissemination of any information or reproduction of any material contained
+ * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
+ * You may not alter or remove any copyright or other notice from copies of this content.
+ */
+
 package org.wso2.carbon.identity.pat.api.rest.commons;
 
 /**
