+++ components/org.wso2.carbon.identity.pat/org.wso2.carbon.identity.pat.api.rest.service.v1/src/gen/java/org/wso2/carbon/identity/pat/api/rest/service/v1/model/TokenMetadataRetrievalResponse.java
@@ -150,7 +150,7 @@ public class TokenMetadataRetrievalResponse  {
         return this;
     }
     
-    @ApiModelProperty(example = "[\"openid\",\"address\"]", required = true, value = "")
+    @ApiModelProperty(example = "[\"internal_application_mgt_view\",\"internal_claim_meta_create\"]", required = true, value = "")
     @JsonProperty("scope")
     @Valid
     @NotNull(message = "Property scope cannot be null.")
