+++ components/org.wso2.carbon.identity.pat/org.wso2.carbon.identity.pat.api.rest.service.v1/src/main/resources/pat-api.yaml
@@ -120,26 +120,16 @@ components:
         - description
         - validity_period
         - scope
-        - id_token_hint
         - client_id
       properties:
         alias:
-          type: string
-          example: "Sample Alias"
+          $ref: '#/components/schemas/Alias'
         description:
-          type: string
-          example: "This is a description for Sample Alias"
+          $ref: '#/components/schemas/Description'
         validity_period:
-          type: integer
-          example: 2000
+          $ref: '#/components/schemas/ValidityPeriod'
         scope:
-          items:
-            type: string
-          type: array
-          example: ["openid", "address"]
-        id_token_hint:
-          type: string
-          example: "eyJ4NXQiOiJNREpsTmpJeE4yRTFPR1psT0dWbU1HUXhPVEZsTXpCbU5tRmpaalEwWTJZd09HWTBOMkkwWXpFNFl6WmpOalJoWW1SbU1tUTBPRGRpTkRoak1HRXdNQSIsImtpZCI6Ik1ESmxOakl4TjJFMU9HWmxPR1ZtTUdReE9URmxNekJtTm1GalpqUTBZMll3T0dZME4ySTBZekU0WXpaak5qUmhZbVJtTW1RME9EZGlORGhqTUdFd01BX1JTMjU2IiwiYWxnIjoiUlMyNTYifQ.eyJhdF9oYXNoIjoiSkNLVW1PUDVLRndHdkdiSzNBSW9YQSIsImF1ZCI6IkpNNnZGQVZYQUFIQTFKd2h3bnY3bjRjTklQMGEiLCJzdWIiOiI3NmQyNmJiZS05MDEwLTRmZTItYmQ3Ni1hNTU5Y2VmMTkyYWEiLCJuYmYiOjE2NDcyNDY4MjMsImF6cCI6IkpNNnZGQVZYQUFIQTFKd2h3bnY3bjRjTklQMGEiLCJhbXIiOlsicGFzc3dvcmQiXSwiaXNzIjoiaHR0cHM6XC9cL2xvY2FsaG9zdDo5NDQzXC9vYXV0aDJcL3Rva2VuIiwiZXhwIjoxNjQ3MjUwNDIzLCJpYXQiOjE2NDcyNDY4MjN9.JfzNT4tx0MsYH5tcHf0Hz4oO_0wnjWro397U6uHwTlONX1w3cdPO9v0hihpcfVq5WOb8VfevGv8EfFwedTZ1fhYg0E1olMjRYh_pajpeOzw8uiej4VuY4dqR5V-KUn_sAVbpkV9Y09VdzSkU5JA2CXhSKy-eLPbvEU8REeHunx3vq7oVz9TQ0EWAnoveomyET6kcuboXWOsD3qBsZaKfr42s5Lr-51H1siaDwuD27REBolUPynrt-YrRqCZ2TFNSR2CfRA0zEaBhikAfsqcBvHQ464iFbXAQcqb8G6B8O6LHoPKVNZlGux-hjoXWAYrUd0axXEN1qDB2JeJiOuv_FA"
+          $ref: '#/components/schemas/Scope'
         client_id:
           type: string
           example: "JM6vFAVXAAHA1Jwhwnv7n4cNIP0a"
@@ -158,25 +148,18 @@ components:
         - scope
       properties:
         token_id:
-          type: string
-          example: "00a12e21-64a1-4b60-9434-3d06b222291c"
+          $ref: '#/components/schemas/TokenId'
         alias:
-          type: string
-          example: "Sample Alias"
+          $ref: '#/components/schemas/Alias'
         description:
-          type: string
-          example: "This is a description for Sample Alias"
+          $ref: '#/components/schemas/Description'
         validity_period:
-          type: integer
-          example: 2000
+          $ref: '#/components/schemas/ValidityPeriod'
         time_created:
           type: string
           example: "2022-03-22 03:28:41"
         scope:
-          items:
-            type: string
-          type: array
-          example: ["openid", "address"]
+          $ref: '#/components/schemas/Scope'
     PATCreationResponse:
       type: object
       required:
@@ -188,26 +171,19 @@ components:
         - validity_period
       properties:
         token_id:
-          type: string
-          example: "00a12e21-64a1-4b60-9434-3d06b222291c"
+          $ref: '#/components/schemas/TokenId'
         access_token:
           type:
             string
           example: "6edb0049-0ea3-3b26-9aa3-152b1891d200"
         alias:
-          type: string
-          example: "Sample Alias"
+          $ref: '#/components/schemas/Alias'
         description:
-          type: string
-          example: "This is a description for Sample Alias"
+          $ref: '#/components/schemas/Description'
         scope:
-          items:
-            type: string
-          type: array
-          example: ["openid", "address"]
+          $ref: '#/components/schemas/Scope'
         validity_period:
-          type: integer
-          example: 2000
+          $ref: '#/components/schemas/ValidityPeriod'
     Error:
       type: object
       required:
@@ -223,6 +199,23 @@ components:
         description:
           type: string
           example: Some Error Description
+    TokenId:
+      type: string
+      example: "00a12e21-64a1-4b60-9434-3d06b222291c"
+    Alias:
+      type: string
+      example: "Sample Alias"
+    Description:
+      type: string
+      example: "This is a description for Sample Alias"
+    ValidityPeriod:
+      type: integer
+      example: 2000
+    Scope:
+      items:
+        type: string
+      type: array
+      example: ["internal_application_mgt_view", "internal_claim_meta_create"]
 
   responses:
     PATListSuccess:
