+++ components/org.wso2.carbon.identity.pat/org.wso2.carbon.identity.pat.core.service/src/main/java/org/wso2/carbon/identity/pat/core/service/PATManagementServiceImpl.java
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.com). All Rights Reserved.
+ *
+ * This software is the property of WSO2 Inc. and its suppliers, if any.
+ * Dissemination of any information or reproduction of any material contained
+ * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
+ * You may not alter or remove any copyright or other notice from copies of this content.
+ */
+
 package org.wso2.carbon.identity.pat.core.service;
 
 import org.apache.commons.logging.Log;
@@ -17,22 +26,21 @@ import org.wso2.carbon.identity.pat.core.service.model.PATCreationReqDTO;
 import org.wso2.carbon.identity.pat.core.service.model.PATCreationRespDTO;
 import org.wso2.carbon.identity.pat.core.service.model.TokenMetadataDTO;
 
-
 import java.util.Arrays;
 import java.util.List;
 
-
-public class PATManagementServiceImpl implements PATManagementService{
+public class PATManagementServiceImpl implements PATManagementService {
 
     private static final Log log = LogFactory.getLog(PATManagementServiceImpl.class);
 
     @Override
     public PATCreationRespDTO issuePAT(PATCreationReqDTO patCreationReqDTO) {
 
-        try{
+        try {
+            String userId = PATUtil.getUserID();
             PATUtil.startSuperTenantFlow();
 
-            OAuth2AccessTokenReqDTO tokenReqDTO = buildAccessTokenReqDTO(patCreationReqDTO);
+            OAuth2AccessTokenReqDTO tokenReqDTO = buildAccessTokenReqDTO(patCreationReqDTO, userId);
             OAuth2AccessTokenRespDTO oauth2AccessTokenResp = PATUtil.getOAuth2Service().issueAccessToken(tokenReqDTO);
 
             if (oauth2AccessTokenResp.getErrorMsg() != null) {
@@ -41,15 +49,15 @@ public class PATManagementServiceImpl implements PATManagementService{
             } else {
                 return getPATCreationResponse(oauth2AccessTokenResp, patCreationReqDTO);
             }
-        }finally {
+        } finally {
             PrivilegedCarbonContext.endTenantFlow();
-
         }
 
     }
 
     @Override
     public TokenMetadataDTO getTokenMetadata(String tokenId) {
+
         PATMgtDAO patMgtDAO = PATDAOFactory.getInstance().getPATMgtDAO();
         TokenMetadataDTO tokenMetadataDTO = patMgtDAO.getTokenMetadata(tokenId);
         tokenMetadataDTO.setScope(patMgtDAO.getTokenScopes(tokenId));
@@ -59,6 +67,7 @@ public class PATManagementServiceImpl implements PATManagementService{
 
     @Override
     public List<TokenMetadataDTO> getTokensMetadata() {
+
         String userId = PATUtil.getUserID();
 
         PATMgtDAO patMgtDAO = PATDAOFactory.getInstance().getPATMgtDAO();
@@ -69,7 +78,8 @@ public class PATManagementServiceImpl implements PATManagementService{
 
     @Override
     public void revokePAT(String tokenId) {
-        try{
+
+        try {
             PATUtil.startSuperTenantFlow();
 
             PATMgtDAO patMgtDAO = PATDAOFactory.getInstance().getPATMgtDAO();
@@ -77,19 +87,20 @@ public class PATManagementServiceImpl implements PATManagementService{
             String clientID = patMgtDAO.getClientIDFromTokenID(tokenId);
 
             OAuthRevocationRequestDTO revokeRequest = buildOAuthRevocationRequest(accessToken, clientID);
-            OAuthRevocationResponseDTO oauthRevokeResp = PATUtil.getOAuth2Service().revokeTokenByOAuthClient(revokeRequest);
+            OAuthRevocationResponseDTO oauthRevokeResp = PATUtil.getOAuth2Service()
+                    .revokeTokenByOAuthClient(revokeRequest);
 
             if (oauthRevokeResp.getErrorMsg() != null) {
                 // TODO: handle error
             }
 
-        }finally {
+        } finally {
             PrivilegedCarbonContext.endTenantFlow();
 
         }
     }
 
-    private OAuth2AccessTokenReqDTO buildAccessTokenReqDTO(PATCreationReqDTO patCreationReqDTO) {
+    private OAuth2AccessTokenReqDTO buildAccessTokenReqDTO(PATCreationReqDTO patCreationReqDTO, String userId) {
 
         OAuth2AccessTokenReqDTO tokenReqDTO = new OAuth2AccessTokenReqDTO();
 
@@ -103,12 +114,13 @@ public class PATManagementServiceImpl implements PATManagementService{
         tokenReqDTO.setClientId(patCreationReqDTO.getClientID());
         tokenReqDTO.setScope(patCreationReqDTO.getScope().toArray(new String[0]));
         // Set all request parameters to the OAuth2AccessTokenReqDTO
-        tokenReqDTO.setRequestParameters(getRequestParameters(patCreationReqDTO));
+        tokenReqDTO.setRequestParameters(getRequestParameters(patCreationReqDTO, userId));
 
         tokenReqDTO.addAuthenticationMethodReference(PATConstants.PAT);
 
         return tokenReqDTO;
     }
+
     private OAuthRevocationRequestDTO buildOAuthRevocationRequest(String accessToken, String clientID) {
 
         OAuthRevocationRequestDTO oAuthRevocationRequestDTO = new OAuthRevocationRequestDTO();
@@ -126,19 +138,23 @@ public class PATManagementServiceImpl implements PATManagementService{
         return oAuthRevocationRequestDTO;
     }
 
+    private RequestParameter[] getRequestParameters(PATCreationReqDTO patCreationReqDTO, String userId) {
 
-    private RequestParameter[] getRequestParameters(PATCreationReqDTO patCreationReqDTO){
         RequestParameter[] requestParameters = new RequestParameter[5];
         requestParameters[0] = new RequestParameter(PATConstants.ALIAS, patCreationReqDTO.getAlias());
         requestParameters[1] = new RequestParameter(PATConstants.DESCRIPTION, patCreationReqDTO.getDescription());
-        requestParameters[2] = new RequestParameter(PATConstants.VALIDITY_PERIOD, String.valueOf(patCreationReqDTO.getValidityPeriod()));
-        requestParameters[3] = new RequestParameter(PATConstants.ID_TOKEN_HINT, patCreationReqDTO.getIdTokenHint());
-        requestParameters[4] = new RequestParameter(PATConstants.SCOPE, patCreationReqDTO.getScope().toArray(new String[0]));
+        requestParameters[2] = new RequestParameter(PATConstants.VALIDITY_PERIOD, String.valueOf(patCreationReqDTO
+                .getValidityPeriod()));
+        requestParameters[3] = new RequestParameter(PATConstants.SCOPE, patCreationReqDTO.getScope()
+                .toArray(new String[0]));
+        requestParameters[4] = new RequestParameter(PATConstants.USER_ID, userId);
 
         return requestParameters;
     }
 
-    private PATCreationRespDTO getPATCreationResponse(OAuth2AccessTokenRespDTO oauth2AccessTokenRespDTO, PATCreationReqDTO patCreationReqDTO){
+    private PATCreationRespDTO getPATCreationResponse(OAuth2AccessTokenRespDTO oauth2AccessTokenRespDTO,
+                                                      PATCreationReqDTO patCreationReqDTO) {
+
         PATCreationRespDTO patCreationRespDTO = new PATCreationRespDTO();
 
         patCreationRespDTO.setTokenId(oauth2AccessTokenRespDTO.getTokenId());
