+++ components/org.wso2.carbon.identity.pat/org.wso2.carbon.identity.pat.core.service/src/main/java/org/wso2/carbon/identity/pat/core/service/common/PATConstants.java
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.com). All Rights Reserved.
+ *
+ * This software is the property of WSO2 Inc. and its suppliers, if any.
+ * Dissemination of any information or reproduction of any material contained
+ * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
+ * You may not alter or remove any copyright or other notice from copies of this content.
+ */
+
 package org.wso2.carbon.identity.pat.core.service.common;
 
 public class PATConstants {
@@ -9,9 +18,9 @@ public class PATConstants {
     public static final String VALIDITY_PERIOD = "validity_period";
     public static final String TIME_CREATED = "time_created";
     public static final String SCOPE = "scope";
-    public static final String ID_TOKEN_HINT = "id_token_hint";
     public static final String TOKEN_SCOPE = "token_scope";
     public static final String ACCESS_TOKEN = "access_token";
     public static final String INBOUND_AUTH_KEY = "inbound_auth_key";
+    public static final String USER_ID = "user_id";
 
 }
