+++ components/org.wso2.carbon.identity.pat/org.wso2.carbon.identity.pat.core.service/src/main/java/org/wso2/carbon/identity/pat/core/service/model/PATCreationReqDTO.java
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.com). All Rights Reserved.
+ *
+ * This software is the property of WSO2 Inc. and its suppliers, if any.
+ * Dissemination of any information or reproduction of any material contained
+ * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
+ * You may not alter or remove any copyright or other notice from copies of this content.
+ */
+
 package org.wso2.carbon.identity.pat.core.service.model;
 
 import java.util.ArrayList;
@@ -8,8 +17,6 @@ public class PATCreationReqDTO {
     private String description;
     private Integer validityPeriod;
     private List<String> scope = new ArrayList<>();
-
-    private String idTokenHint;
     private String clientID;
 
     public String getClientID() {
@@ -52,11 +59,4 @@ public class PATCreationReqDTO {
         this.scope = scope;
     }
 
-    public String getIdTokenHint() {
-        return idTokenHint;
-    }
-
-    public void setIdTokenHint(String idTokenHint) {
-        this.idTokenHint = idTokenHint;
-    }
 }
